+++ testing/tests/TestUserBasic.php
@@ -2,7 +2,7 @@
 
 use Fr\LS;
 
-class TestUserBasic extends PHPUnit_Framework_TestCase
+class TestUserBasic extends \PHPUnit\Framework\TestCase
 {
     private static $pdo = null;